build: pin the language standard to C++17 (#1360)
aMule never set a language standard, so every compiler picked its own. Measured with `__cplusplus` on each toolchain we build with:

| Toolchain | `__cplusplus` | Standard |
|---|---|---|
| AppleClang 21 (macOS) | `201402` | **C++14** |
| g++ 15 (Ubuntu) | `201703` | C++17 |
| clang++ 21 (Ubuntu) | `201703` | C++17 |
| clang 22 (mingw-w64) | `201703` | C++17 |

The same source has been a different language depending on who built it, with macOS the odd one out.

## Why this is worth fixing

A `static constexpr` member is implicitly inline from C++17. Before that, odr-using one (binding a reference, which `std::min()` does) needs an out-of-line definition, and a header-only class has nowhere to put it:

| Declaration | C++14 | C++17 |
|---|---|---|
| `static const size_t k = ...;` | link error | link error |
| `static constexpr size_t k = ...;` | **link error** | links |
| `static constexpr size_t Fn() { ... }` | links | links |

Code in that middle row links on three platforms and fails on the fourth, as an undefined symbol in the *caller's* translation unit rather than anything pointing at the declaration. CI catches it, since macOS is in the matrix, but the message tells you almost nothing.

It has cost us twice already: the vendored libutp in #1353 compiled with no `-std` at all, its C++17 `if`-init statements surviving only as a compiler extension, and #1359 works around the odr-use rule with a `constexpr` function because a member could not be relied on.

## Why 17 and not 14

Three of the four toolchains already default to C++17, so pinning it moves only macOS while pinning 14 would move three platforms backwards. C++17 is nine years old, fully supported since Clang 7, GCC 9 and MSVC 15.7. Our CMake floor of 3.10 is well past the 3.8 that added `CXX_STANDARD 17`.

## What changed

**The pin**, in the top-level `CMakeLists.txt`: `CMAKE_CXX_STANDARD 17` with `CXX_STANDARD_REQUIRED ON`.

**Two suppressions the pin requires.** Boost.Asio's execution headers define constexpr statics out of line, which C++17 makes redundant and deprecates, tripping the `-Werror=deprecated` gate. `LibSocketAsio.cpp` and `webapi/HttpServer.cpp` already wrapped their asio includes for the sibling diagnostic and now cover this one too. Under the old C++14 default those definitions were still *required*, which is why this appears only now. `NetworkAddressAsio.h` needs nothing: `ip/address.hpp` does not reach the execution headers, measured at zero errors.

**Five workarounds the split forced, now removed**, along with the comments justifying them:

| Site | Before | After |
|---|---|---|
| `AICHHashList.cpp/.h` | `uint16_t(INVALID_INDEX)` cast, to avoid odr-using a `constexpr` member | bound directly |
| `PublicIPv6Corroboration.h` | discarded `emplace_back()`'s return, "the unit-test targets take clang's default of C++14" | uses the returned reference |
| `webapi/HttpServer.cpp` | `&out[0]`, since "`data()` is const-qualified pre-C++17" | `out.data()` |
| `webapi/Api.cpp` | `boost::optional` justified partly by "the tree builds as C++14" | unchanged in code; the false clause is gone, house style stands on its own |
| `cmake/libutp.cmake` | comment: "aMule sets `CMAKE_CXX_STANDARD` nowhere" | corrected. Its own `CXX_STANDARD 17` stays, because that is upstream's requirement rather than ours to lend |

**A refactor, in its own commit so it can be dropped.** Ten files each spelled out the same pragma block around their third-party includes, in two variants: three flags for Crypto++, two for Boost.Asio. The pin would have made it eleven. `WarningsPush_CryptoPP.h` and `WarningsPush_Asio.h` now hold one copy of each set, closed by a shared `WarningsPop.h`. All three are deliberately unguarded, since a translation unit opens and closes a block per include group.

A push/pop pair rather than one wrapper that includes everything, because the sites include narrow subsets on purpose: `CryptoPP_Inc.h` pulls wider than libwebcommon wants, and #1350 split `AddressFamilyPolicyAsio.h` out specifically to keep asio out of most of `src/`. Routing them through a common include would undo both.

They are included by path relative to each file rather than by name, because the `webcommon` target compiles with only `-I src/libwebcommon` and a bare name resolves nowhere there. A quoted include searches the includer's directory first, so the relative form works whatever a target's `-I` list happens to be.

## Verification

| Build | Result |
|---|---|
| macOS, `BUILD_EVERYTHING` + testing + `ENABLE_UTP`, all 12 binaries | 0 errors, **61/61**, one pre-existing `-Wswitch` |
| Ubuntu gcc 15, monolithic + remotegui + daemon | 0 errors, **60/60** |
| `KadAICHHashListTest`, covering the file the cast was removed from | passes |

Compile lines carry `-std=gnu++17`. clang-format 18 clean on every changed file, `git diff --check` clean.

## Effect on the lint gates

| Tier | C++14 | C++17 |
|---|---|---|
| Tier-1, whole-tree gate | 6 findings | 6, unchanged |
| Tier-2, changed lines | 594 | 688 across six files |

Tier-1 does not move. Tier-2's increase is **not new check categories** (diffing the check names shows none appears only under C++17), just more instances of checks already enabled, and it only ever gates changed lines.

## Side effect worth having

Under C++14 macOS emitted roughly ninety `-Wc++17-extensions` warnings from Crypto++ headers using C++17 `static_assert` without a message. Those disappear by definition. The full build now produces one warning in total.

+++ src/kademlia/kademlia/AICHHashList.cpp
@@ -85,10 +85,7 @@ const CKadAICHHash &CKadAICHHashList::GetHashAt(uint16_t index) const
 
 std::vector<uint16_t> CKadAICHHashList::BuildCompactionMap() const
 {
-	// uint16_t(...) rather than INVALID_INDEX: the fill constructor takes a
-	// const reference, which would ODR-use the member and need an
-	// out-of-line definition -- ill-formed for a constexpr member in C++17.
-	std::vector<uint16_t> map(m_hashes.size(), uint16_t(INVALID_INDEX));
+	std::vector<uint16_t> map(m_hashes.size(), INVALID_INDEX);
 	uint16_t next = 0;
 	for (size_t i = 0; i < m_hashes.size(); ++i) {
 		if (m_popularity[i] > 0) {
